+++ H3Status/Plugin.cs
@@ -1,45 +1,54 @@
 using BepInEx;
 using BepInEx.Logging;
 using HarmonyLib;
-using WebSocketSharp.Server;
 
-namespace H3Status;
-
-[BepInProcess("h3vr.exe")]
-[BepInPlugin(Guid, Name, Version)]
-public class Plugin : BaseUnityPlugin
+namespace H3Status
 {
-    public const string Guid = "xyz.bacur.plugins.h3status";
-    public const string Name = "H3Status";
-    public const string Version = "0.3.0";
-
-
-    internal static new ManualLogSource Logger;
-    private static Harmony _harmony;
-
-    public static HttpServer server;
-    public static int port = 9504;
-
-    private void Awake()
+    [BepInProcess("h3vr.exe")]
+    [BepInPlugin(Guid, Name, Version)]
+    public class Plugin : BaseUnityPlugin
     {
-        Logger = base.Logger;
-        server = new HttpServer(port);
-        server.AddWebSocketService<Server.ServerBehavior>("/");
-        server.Start();
+        public const string Guid = "xyz.bacur.plugins.h3status";
+        public const string Name = "H3Status";
+        public const string Version = "0.5.0";
 
-        Logger.LogInfo($"Server started on port {port}");
+        internal static new ManualLogSource Logger;
+        internal static Harmony Patcher;
 
-        _harmony = new Harmony(Guid);
-        _harmony.PatchAll(typeof(Patches.SceneHandler));
-        _harmony.PatchAll(typeof(Patches.TNHScoreHandler));
-        _harmony.PatchAll(typeof(Patches.TNHPhaseHandler));
-        _harmony.PatchAll(typeof(Patches.PlayerHealthHandler));
-        _harmony.PatchAll(typeof(Patches.WeaponAmmoHandler));
-    }
+        protected void Awake()
+        {
+            Logger = base.Logger;
+            Patcher = new Harmony(Guid);
+            H3Status.Config.Init(Config);
 
-    private void OnDestroy()
-    {
-        server.Stop();
-        _harmony.UnpatchSelf();
+            if (!H3Status.Config.ServerEnabled.Value)
+            {
+                Logger.LogWarning("Server disabled in config");
+                return;
+            }
+
+            Start();
+        }
+
+        protected void OnDestroy()
+        {
+            Stop();
+        }
+
+        internal static void Start()
+        {
+            Server.Start(H3Status.Config.ServerPort.Value);
+            Patcher?.PatchAll(typeof(Patches.SceneHandler));
+            Patcher?.PatchAll(typeof(Patches.TNHScoreHandler));
+            Patcher?.PatchAll(typeof(Patches.TNHPhaseHandler));
+            Patcher?.PatchAll(typeof(Patches.PlayerHealthHandler));
+            Patcher?.PatchAll(typeof(Patches.WeaponAmmoHandler));
+        }
+
+        internal static void Stop()
+        {
+            Server.Stop();
+            Patcher?.UnpatchSelf();
+        }
     }
 }