+++ Overlays/TNHEncryptionTimer/main.js
@@ -10,6 +10,7 @@ function connect() {
   return ws;
 }
 
+const pointsPerSecond = 10 * 10;
 let scoreMultiplier = 10;
 let encryptionType;
 let encryptionCount;
@@ -35,10 +36,12 @@ function onMessage(e) {
       hideOverlay();
       break;
 
+    case "tnhLevelEvent":
     case "TNHLevelEvent":
       scoreMultiplier = event.status.scoreMultiplier;
       break;
 
+    case "tnhEncryptionDestroyed":
     case "TNHEncryptionDestroyed":
       // Cascading turns into multiple full encryptions,
       // so the counter would go into the negatives here
@@ -53,6 +56,7 @@ function onMessage(e) {
       drawTickMark();
       break;
 
+    case "tnhHoldPhaseEvent":
     case "TNHHoldPhaseEvent":
       if (event.status.phase == "Analyzing") {
         encryptionType = event.status.encryptionType;
@@ -64,7 +68,10 @@ function onMessage(e) {
           `icons/${encryptionType}.webp`;
 
         clockText.setValue(0, 500);
-        scoreText.setValue(countdown.duration * 50 * scoreMultiplier, 500);
+        scoreText.setValue(
+          countdown.duration * pointsPerSecond * scoreMultiplier,
+          500,
+        );
         scoreLostText.setValue(0, 500);
 
         showOverlay();
@@ -83,6 +90,7 @@ function onMessage(e) {
       }
       break;
 
+    case "tnhPhaseEvent":
     case "TNHPhaseEvent":
       if (event.status.phase == "Take" || event.status.phase == "Completed") {
         hideOverlay();
@@ -156,8 +164,8 @@ class Countdown {
 
     const progress = this.value / this.duration;
     const clock = this.duration - this.value;
-    const score = this.value * 50 * scoreMultiplier;
-    const scoreLost = 6000 * scoreMultiplier - score;
+    const score = this.value * pointsPerSecond * scoreMultiplier;
+    const scoreLost = 12000 * scoreMultiplier - score;
 
     if (this.value <= 60) timerBar.setColor("#f44");
     timerBar.setValue((progress % 0.5) * 2);